requirements.txt
pydantic>=2.0,<3
fastapi>=0.100,<1
uvicorn[standard]>=0.22,<1
httpx>=0.25,<1
openai>=1.0,<2
python-dotenv>=1.0,<2
youtube-transcript-api>=1.2,<2
